Add two xapps compatible with RIC Taap project     - Energy Saving with cell utilization xapp     - Energy Saving with load balancing xapp     - Create custom rc_ctrl_service_style_300 for energy state.

+++ src/sm/kpm_sm/kpm_sm_v02.01/ie/kpm_data_ie/kpm_ric_info/kpm_ric_ind_hdr_frm_1.h
@@ -11,7 +11,7 @@ extern "C" {
 // 8.2.1.3.1  E2SM-KPM Indication Header Format 1
 
 typedef struct {
-    uint32_t collectStartTime;  // 8.3.12
+    uint64_t collectStartTime;  // 8.3.12
     byte_array_t *fileformat_version;  /* OPTIONAL */
     byte_array_t *sender_name;         /* OPTIONAL */
     byte_array_t *sender_type;         /* OPTIONAL */
